Feat: updated subscription lifecycle
