+++ apps/controllo-ore-x/src/app/utils/subscriptions_lifecycle.ts
@@ -1,0 +1,30 @@
+import { Subscription } from 'rxjs';
+
+export interface SubscriptionsLifecycle {
+  subscriptionsList: Subscription[];
+
+  ngOnInit(): void;
+  ngOnDestroy(): void;
+
+  /**
+   * Subscribes to the provided subscription and adds it to the list.
+   */
+  _setSubscriptions(): void;
+
+  /**
+   * Unsubscribes from all the subscriptions in the list.
+   * Should use @see {@link completeSubscriptions}
+   */
+  _completeSubscriptions: typeof completeSubscriptions;
+}
+
+/**
+ * Iterate through the subscriptionsList and unsubscribe from each subscription in the list.
+ */
+export const completeSubscriptions = (
+  subscriptionsList: Subscription[],
+): void => {
+  for (const subscription of subscriptionsList) {
+    subscription.unsubscribe();
+  }
+};
