--- libs/controllo-ore-x-interfaces/src/index.ts
@@ -1,1 +1,0 @@
-export * from './subscriptions-lifecycle';
